brain_tumor_classification/requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
pillow>=8.3.0
tqdm>=4.62.0

# Jupyter notebook support
jupyter>=1.0.0
ipykernel>=6.0.0

# Additional utilities
opencv-python>=4.5.0
albumentations>=1.3.0
tensorboard>=2.8.0

# For Kaggle dataset download (optional)
kaggle>=1.5.0

# For uncertainty estimation
scipy>=1.7.0

# For visualization and analysis
plotly>=5.0.0